调试分类数据并处理基础架构

+++ shop-mbg/shop-mbg-product/src/main/java/com/sunny/product/model/ProductNavCategory.java
@@ -32,7 +32,7 @@ public class ProductNavCategory implements Serializable {
      *
      * @mbg.generated
      */
-    private Byte status;
+    private int status;
 
     private Date createTime;
 
@@ -100,11 +100,11 @@ public class ProductNavCategory implements Serializable {
         this.description = description;
     }
 
-    public Byte getStatus() {
+    public int getStatus() {
         return status;
     }
 
-    public void setStatus(Byte status) {
+    public void setStatus(int status) {
         this.status = status;
     }
